// frontend/src/app/openapi/model/badge-output.ts
/*
 * SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
 * SPDX-License-Identifier: Apache-2.0
 *
 * Capella Collaboration Manager API
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit the class manually.
 + To generate a new version, run `make openapi` in the root directory of this repository.
 */

import { Text } from './text';
import { BadgeVariant } from './badge-variant';


export interface BadgeOutput { 
    /**
     * Show a badge with the current environment.
     */
    show: boolean;
    /**
     * Color of the badge.
     */
    variant: BadgeVariant;
    text: Text;
}
export namespace BadgeOutput {
}